Entrega del producto del Sprint 1 testeado sobre Develop

+++ Analizar/frontend/analizar/src/app/pages/alertas/alertas.component.ts
@@ -5,6 +5,7 @@ import { AlertasService } from 'src/app/service/alertas.service';
 import { MedidoresService } from 'src/app/service/medidores.service';
 
 import { Alert } from './Alert';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -23,16 +24,16 @@ constructor(private router: Router, private alertaService: AlertasService, priva
  medidores: any[] = [];
  alertasForm!: FormGroup;
  loginError: string = '';
- 
+
  ngOnInit(): void{
   this.alertasForm = this.initForm()
   this.getAlerts()
   this.getMedidoresByUser()
-  
+
  }
  getAlerts(){
   this.alertaService.getAlertas().subscribe((data: any) => {
-    this.alertas = data   
+    this.alertas = data
   })
  }
  //Traer medidores para poder seleccionar para que medidor hacer la alerta
@@ -40,14 +41,14 @@ constructor(private router: Router, private alertaService: AlertasService, priva
   this.medidorService.getMedidores().subscribe((data: any) => {
     this.medidores = data
     //console.log(data);
-    
+
   })
  }
    //Validaciones para los campos
  saveAlert(){
-  const valor = this.alertasForm.value.valor; 
+  const valor = this.alertasForm.value.valor;
   const medidor = this.alertasForm.value.medidor;
-  const fechaAlta = this.alertasForm.get('fechaAlta')?.value; 
+  const fechaAlta = this.alertasForm.get('fechaAlta')?.value;
   if(this.alertasForm.valid){
   this.alertaService.addAlertas(valor, medidor, fechaAlta).subscribe((alert: any) => {
     //console.log('Alerta agregada con éxito:', alert);
@@ -67,7 +68,7 @@ constructor(private router: Router, private alertaService: AlertasService, priva
     alert.setAlert = this.setAlert;
     alert.typeAlert = this.typeAlert;
     console.log(this.setAlert);
-    console.log(this.typeAlert);  
+    console.log(this.typeAlert);
     this.alerts.push(alert);
     this.setAlert = 0;
     this.typeAlert = '';
@@ -83,13 +84,35 @@ constructor(private router: Router, private alertaService: AlertasService, priva
  }
 
  removeAlert(id:number){  
-  this.alertaService.removeAlertas(id).subscribe((alert) => {
-    //console.log('Alerta eliminada con éxito:', alert);
-    this.closeModal();
-    this.getAlerts()
-  }, (error: any) => {
-    console.error('Hubo un error al eliminar la alerta', error);
-  })
+
+  Swal.fire({
+    title: "Estás seguro de eliminar la alerta?",
+    text: "No podrás revertir esto!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Sí, eliminar!"
+  }).then((result) => {
+    if (result.isConfirmed) {
+
+      this.alertaService.removeAlertas(id).subscribe((alert) => {
+        //console.log('Alerta eliminada con éxito:', alert);
+        this.closeModal();
+        this.getAlerts()
+      }, (error: any) => {
+        console.error('Hubo un error al eliminar la alerta', error);
+      })
+
+      Swal.fire({
+        title: "Eliminada!",
+        text: "La alerta ha sido eliminada con éxito",
+        icon: "success"
+      });
+    }
+  });
+
+  
 
  }
 
@@ -98,8 +121,8 @@ constructor(private router: Router, private alertaService: AlertasService, priva
     const modal = document.getElementById('addAlert');
     let contenedorAlertas = document.getElementById('contenedor-alertas');
     if(modal != null) {
-      modal.style.display ='flex'; 
-      this.alertasForm.reset();     
+      modal.style.display ='flex';
+      this.alertasForm.reset();
     }
     if(contenedorAlertas != null) {
       contenedorAlertas.style.display ='none';
@@ -110,7 +133,7 @@ constructor(private router: Router, private alertaService: AlertasService, priva
     const editModal = document.getElementById('editAlert');
     let contenedorAlertas = document.getElementById('contenedor-alertas');
     if(editModal != null) {
-      editModal.style.display ='flex';      
+      editModal.style.display ='flex';
     }
     if(contenedorAlertas != null) {
       contenedorAlertas.style.display ='none';
@@ -133,6 +156,16 @@ constructor(private router: Router, private alertaService: AlertasService, priva
       medidor: ['', [Validators.required, Validators.pattern('^[0-9]+$')]],
       fechaAlta: [''],
     })
+  }
+
+  handleEnterKey(event: Event): void {
+    if (event instanceof KeyboardEvent && event.key === 'Enter') {
+      const targetElement = event.target as HTMLElement;
+      if (targetElement.tagName === 'A' && targetElement.classList.contains('nav-link')) {
+        const linkText = targetElement.textContent?.trim();
+        console.log(`Se presionó Enter en el enlace: ${linkText}`);
+      }
+    }
   }
 
 }
