+++ Analizar/frontend/analizar/src/app/pages/medidores/medidores.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup, NgModel, Validators } from '@angular/forms';
 import { Router } from '@angular/router';
 import { MedidoresService } from 'src/app/service/medidores.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-medidores',
@@ -16,22 +17,22 @@ constructor(private router: Router, private medidorService: MedidoresService, pr
  medidores: any[] = [];
  medidoresForm!: FormGroup;
  loginError: string = '';
- 
+
  ngOnInit(): void{
   this.medidoresForm = this.initForm()
   this.getMedidores()
-  
+
  }
  getMedidores(){
   this.medidorService.getMedidores().subscribe((data: any) => {
-    this.medidores = data   
+    this.medidores = data
   })
  }
    //Validaciones para los campos
  saveMedidor(){
-  const nombre = this.medidoresForm.value.nombre; 
+  const nombre = this.medidoresForm.value.nombre;
   const detalle = this.medidoresForm.value.detalle;
-  const identificador = this.medidoresForm.get('identificador')?.value; 
+  const identificador = this.medidoresForm.get('identificador')?.value;
   let userId = Number(localStorage.getItem('userId')!);
   console.log(nombre, detalle, identificador, userId);
   if(this.medidoresForm.valid){
@@ -48,13 +49,35 @@ constructor(private router: Router, private medidorService: MedidoresService, pr
   }
  }
  removeAlert(id:number){  
-  this.medidorService.removeMedidor(id).subscribe((medidor) => {
-    console.log('Alerta eliminada con éxito:', medidor);
-    this.closeModal();
-    this.getMedidores()
-  }, (error: any) => {
-    console.error('Hubo un error al eliminar la alerta', error);
-  })
+
+  Swal.fire({
+    title: "Estás seguro de eliminar el medidor?",
+    text: "No podrás revertir esto!",
+    icon: "warning",
+    showCancelButton: true,
+    confirmButtonColor: "#3085d6",
+    cancelButtonColor: "#d33",
+    confirmButtonText: "Sí, eliminar!"
+  }).then((result) => {
+    if (result.isConfirmed) {
+
+      this.medidorService.removeMedidor(id).subscribe((medidor) => {
+        console.log('Medidor eliminado con éxito:', medidor);
+        this.closeModal();
+        this.getMedidores()
+      }, (error: any) => {
+        console.error('Hubo un error al eliminar el medidor', error);
+      })
+
+      Swal.fire({
+        title: "Eliminado!",
+        text: "El medidor ha sido eliminado con éxito",
+        icon: "success"
+      });
+    }
+  });
+
+  
 
  }
 
@@ -63,8 +86,8 @@ constructor(private router: Router, private medidorService: MedidoresService, pr
     const modal = document.getElementById('addAlert');
     let contenedorAlertas = document.getElementById('contenedor-alertas');
     if(modal != null) {
-      modal.style.display ='flex'; 
-      this.medidoresForm.reset();     
+      modal.style.display ='flex';
+      this.medidoresForm.reset();
     }
     if(contenedorAlertas != null) {
       contenedorAlertas.style.display ='none';
@@ -88,5 +111,16 @@ constructor(private router: Router, private medidorService: MedidoresService, pr
       identificador: [''],
     })
   }
+
+  handleEnterKey(event: Event): void {
+    if (event instanceof KeyboardEvent && event.key === 'Enter') {
+      const targetElement = event.target as HTMLElement;
+      if (targetElement.tagName === 'A' && targetElement.classList.contains('nav-link')) {
+        const linkText = targetElement.textContent?.trim();
+        console.log(`Se presionó Enter en el enlace: ${linkText}`);
+      }
+    }
+  }
+
 
 }
