+++ Analizar/frontend/analizar/src/app/pages/planes/planes.component.ts
@@ -21,19 +21,41 @@ export class PlanesComponent implements OnInit {
       //console.log(servicios);
       this.servicios = servicios;
     }, (error: any) => {
-      console.error('Hubo un error al obtener los servicios', error);      
+      console.error('Hubo un error al obtener los servicios', error);
     })
   }
-  isAdmin(): boolean {  
+  isAdmin(): boolean {
     return this.authService.getIsAdmin();
   }
   removeProducto(id: number){
-    this.productosService.removeProducto(id).subscribe((product: any) => {
-      console.log('Producto eliminado con exito:', product);
-      this.getServicios()
-    }, (error: any) => {
-      console.error('Hubo un error al eliminar el producto', error);      
-    })
+
+    Swal.fire({
+      title: "Estás seguro de eliminar el plan?",
+      text: "No podrás revertir esto!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Sí, eliminar!"
+    }).then((result) => {
+      if (result.isConfirmed) {
+  
+        
+        this.productosService.removeProducto(id).subscribe((product: any) => {
+          console.log('Plan eliminado con exito:', product);
+          this.getServicios()
+        }, (error: any) => {
+          console.error('Hubo un error al eliminar el plan', error);
+        })
+  
+        Swal.fire({
+          title: "Eliminado!",
+          text: "El plan ha sido eliminado con éxito",
+          icon: "success"
+        });
+      }
+    });
+
   }
 
   agregarCarrito(producto:any, tipoProducto:string){
@@ -52,17 +74,28 @@ export class PlanesComponent implements OnInit {
       if(!seAgregoElemento){
         carritoActual!.push({producto,"cantidad":1, tipoProducto})
       }
-      
+
       localStorage.setItem('mi-carrito',JSON.stringify(carritoActual))
     }else{
       let carritoActual = JSON.stringify([{producto,"cantidad":1, tipoProducto}])
       localStorage.setItem('mi-carrito',carritoActual)
     }
-    
+
     Swal.fire({
       icon: 'success',
       title: 'Éxito!',
       text: 'Se agregó el producto al carrito!'
     })
   }
+
+  handleEnterKey(event: Event): void {
+    if (event instanceof KeyboardEvent && event.key === 'Enter') {
+      const targetElement = event.target as HTMLElement;
+      if (targetElement.tagName === 'A' && targetElement.classList.contains('nav-link')) {
+        const linkText = targetElement.textContent?.trim();
+        console.log(`Se presionó Enter en el enlace: ${linkText}`);
+      }
+    }
+  }
+
 }
